+++ src/components/user-action/Checkbox.tsx
@@ -79,7 +79,7 @@ const Checkbox = ({
         className={clsx(usedSizeClass, `items-center border-2 rounded outline-none `, {
           'text-disabled-text border-disabled-outline bg-disabled-background cursor-not-allowed': disabled,
           'focus:border-primary group-hover:border-primary ': !disabled,
-          'bg-surface': !disabled && !checked,
+          'bg-input-background': !disabled && !checked,
           'bg-primary/30 border-primary text-primary':  !disabled && checked === true || checked === 'indeterminate',
         }, className)}
       >