+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.1.22] - 2025-07-16
+
+### Added
+- Added `coloredHoverBackground` option to `TextButton`
+- Added colors for input elements to css instead of using surface
+
+### Changed
+- Change `Avatar` component to allow for name displays
+- Change color for `Modal`s and `Menu`s
+
+### Fixed
+- Fixed `Menu` and `Overlay` z-Indexes
+
 ## [0.1.21] - 2025-07-14
 
 ### Added